+++ test_classes.py
@@ -4,25 +4,27 @@ from unit_converter.api import Converter
 converter = Converter()
 
 
-# converter.groups()
-
-# message = converter.types("length", "a")
+# message = converter.groups()
 # print(message)
 
-# message = converter.history(limit=2, another="a")
+
+# message = converter.types("length")
 # print(message)
 
-# message = converter.convert("length", "m 10")
-# print(str(message))
+# message = converter.history()
+# print(message)
 
-#converter.manage_group("metro", "remove")
-#print(f"Result: {result}")
+# message = converter.convert("length", "meters feet 5", print_message=True)
+# print(message)
 
-# result = converter.manage_type("temperature", "add new_type 1 1")
-# print(f"Result: {result}")
+#converter.manage_group("new_group", "remove", print_message=True)
+# print(message)
 
-result = converter.change_base("length", "meters")
-print(f"Result: {result}")
+# message = converter.manage_type("length", "remove new_type")
+# print(message)
 
-# result = converter.aliases("length", "remove meters mtr", print_message=True)
-# print(f"Result: {result}")
+# message = converter.aliases("length", "add meters mtr")
+# print(message)
+
+# message = converter.change_base("length", "meters")
+# print(message)